+++ Herbarium/src/BlockEntity/BEShrubBerryBush.cs
@@ -1,5 +1,4 @@
 using Vintagestory.API.Client;
-using Vintagestory.API.Common;
 
 namespace herbarium 
 {
